Fixed warning
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@6653 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/CoolMenu.cpp
@@ -699,7 +699,7 @@ public:
 			m_pID[ i ] = GetNext( pos )->m_pLastId;
 		}
 
-		return SUCCEEDED( m_pFolder->GetUIObjectOf( hWnd, GetCount(),
+		return SUCCEEDED( m_pFolder->GetUIObjectOf( hWnd, (UINT)GetCount(),
 			m_pID, IID_IContextMenu, NULL, ppContextMenu ) );
 	}
